Made use of iLoader to load textures
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@35865 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/mesh/persist/assimpldr/importanimesh.cpp
@@ -156,7 +156,7 @@ CS_PLUGIN_NAMESPACE_BEGIN(AssimpLoader)
 	(&animeshData, node, CS::Animation::InvalidBoneID, aiMatrix4x4 ());
 
       printf ("Skeleton: %i bones:\n%s\n",
-	      animeshData.factory->GetSkeletonFactory()->GetTopBoneID (),
+	      (int) animeshData.factory->GetSkeletonFactory()->GetTopBoneID (),
 	      animeshData.factory->GetSkeletonFactory()->Description().GetData());
       
       if (animeshData.factory->GetVertexCount () > 0)
